Moved files, created a package and refactored/commented files
