+++ app/src/main/java/com/github/displace/sdp2022/database/DatabaseConstants.kt
@@ -1,0 +1,5 @@
+package com.github.displace.sdp2022.database
+
+object DatabaseConstants {
+    const val DB_URL = "https://displace-dd51e-default-rtdb.europe-west1.firebasedatabase.app/"
+}
