+++ app/src/main/java/com/github/displace/sdp2022/settings/SettingsConstants.kt
@@ -1,0 +1,6 @@
+package com.github.displace.sdp2022.settings
+
+object SettingsConstants {
+    const val PURPLE_THEME_KEY = "purple"
+    const val GREEN_THEME_KEY = "green"
+}
