+++ app/src/main/java/com/github/displace/sdp2022/settings/SettingsFragment.kt
@@ -1,9 +1,17 @@
+package com.github.displace.sdp2022.settings
+
 import android.os.Bundle
 import androidx.preference.PreferenceFragmentCompat
 import androidx.preference.PreferenceManager
 import com.github.displace.sdp2022.R
 
-
+/**
+ * Settings fragment
+ *
+ * Just a very simple settings fragment, that will be used to display the settings.
+ *
+ * @author blecoeur
+ */
 class SettingsFragment : PreferenceFragmentCompat() {
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         setPreferencesFromResource(R.xml.preferences, rootKey)
